edit logic userSlice
