+++ src/store/index.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import userReducer from './slices/registeredUserSlice';
+import userReducer from './slices/userSlice.js';
 
 const store = configureStore({
   reducer: {
